# docs/swagger.yaml
basePath: /
definitions:
...
  api.CalculateLinesRequest:
    properties:
      input:
        type: string
    type: object
...
  api.ImageMetadata:
    properties:
      color:
        type: string
      color_counts:
        additionalProperties:
          type: integer
        type: object
      description:
        type: string
      dimensions:
        type: string
      icc_description:
        type: string
      icc_profile:
        type: string
      jfif:
        $ref: '#/definitions/api.JFIFMetadata'
    type: object
  api.InversionResult:
    properties:
      d:
        type: integer
      g_d:
        type: number
      mu_n_d:
        type: integer
      term:
        type: number
    type: object
...
  api.JFIFMetadata:
    properties:
      units:
        type: string
      version:
        type: string
      x_density:
        type: integer
      y_density:
        type: integer
    type: object
...
  api.MobiusDirectRequest:
    properties:
      numbers:
        description: Comma separated or single number
        type: string
    type: object
  api.MobiusDirectResponse:
    properties:
      mu:
        type: integer
      "n":
        type: integer
      reason:
        type: string
    type: object
  api.MobiusDivisorRequest:
    properties:
      "n":
        type: integer
    type: object
  api.MobiusDivisorResponse:
    properties:
      divisors:
        items:
          type: integer
        type: array
      inversion_details:
        items:
          $ref: '#/definitions/api.InversionResult'
        type: array
      "n":
        type: integer
      sum_mu_d:
        type: integer
      total_inversion:
        type: number
    type: object
  api.MobiusErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.MobiusMaskRequest:
    properties:
      data:
        items:
          type: number
        type: array
    type: object
...
host: localhost:3301
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Quantix Math API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Quantix Math API
  version: "1.0"
paths:
...
  /api/liber-primus/metadata/{id}:
    get:
      description: Returns metadata (dimensions, colors, JFIF info) for a Liber Primus
        page image
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ImageMetadata'
      summary: Get Liber Primus page metadata
      tags:
      - 3301 Tools
  /api/mobius/direct:
    post:
      consumes:
      - application/json
      description: Returns the Mobius function value (mu) for a list of numbers
      parameters:
      - description: Mobius Direct Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.MobiusDirectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.MobiusDirectResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.MobiusErrorResponse'
      summary: Calculate Mobius function for numbers
      tags:
      - Math
  /api/mobius/divisor:
    post:
      consumes:
      - application/json
      description: Returns the divisor sum and Mobius inversion details for a number
        N
      parameters:
      - description: Mobius Divisor Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.MobiusDivisorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MobiusDivisorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.MobiusErrorResponse'
      summary: Calculate Mobius divisor sum and inversion
      tags:
      - Math
  /api/mobius/mask:
    post:
      consumes:
      - application/json
      description: Applies the Mobius sequential index masking to the provided data
      parameters:
      - description: Mobius Mask Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.MobiusMaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              format: float64
              type: number
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.MobiusErrorResponse'
      summary: Apply Mobius masking to a data sequence
      tags:
      - Math
...
  /api/runer/calculate-lines:
    post:
      consumes:
      - application/json
      description: Returns the calculated rune lines for the given input string
      parameters:
      - description: Calculate Lines Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CalculateLinesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Invalid request body
          schema:
            type: string
      summary: Calculate rune lines from input
      tags:
      - 3301 Tools
...
swagger: "2.0"